+++ projects/triangular-engine/pmndrs/components/index.ts
@@ -1,0 +1,3 @@
+export * from './billboard/billboard.component';
+export * from './sparkles/sparkles.component';
+export * from './pmndrs-components.module';