+++ projects/triangular-engine/pmndrs/pmndrs.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { PmndrsComponentsModule } from './components';
+
+const importExport = [PmndrsComponentsModule] as const;
+
+@NgModule({
+  imports: [...importExport],
+  exports: [...importExport],
+})
+export class PmndrsModule {}